[Behat] Run managing catalog promotion scenarios in separation
This change was the simplest scenario for single scecnario failing. This scenario was failing only in UI context(while working fine in API). The problem didn't exist when scenarios where run locally or just in separate step of GitHub Action. Instead of commiting more time for the investigation, I've decided to split the execution.